Added property to V1
