--- src/OmniSharp.Abstractions/Models/v2/TypeLookUp/TypeLookUpRequest.cs
@@ -1,11 +1,0 @@
-﻿using OmniSharp.Mef;
-
-namespace OmniSharp.Models.v2.TypeLookUp
-{
-    [OmniSharpEndpoint(OmniSharpEndpoints.V2.TypeLookup, typeof(TypeLookupRequest), typeof(TypeLookupResponse))]
-    public class TypeLookupRequest : Request
-    {
-        public bool IncludeDocumentation { get; set; }
-    }
-}
-
