--- src/OmniSharp.Abstractions/Models/v2/TypeLookUp/TypeLookUpResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace OmniSharp.Models.v2.TypeLookUp
-{
-    public class TypeLookupResponse
-    {
-            public string Type { get; set; }
-            public DocumentationComment DocComment { get; set; }
-    }   
-}
